# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1",
      "description": "local server"
    }
  ],
  "paths": {
...
    "/foods/create-food": {
      "post": {
        "tags": ["default"],
        "summary": "CreateFood",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "string",
                    "example": "{\n  \"foodName\": \"Food Name\",\n  \"storeId\": \"65c2b09b0546d4998ffa4daf\",\n  \"image\": \"image.jpg\",\n  \"price\": 10.99,\n  \"regularPrice\": 12.99,\n  \"offer\": true,\n  \"offerPercentage\": 10,\n  \"offerPrice\": 11.69,\n  \"category\": \"Category\",\n  \"video\": \"video.mp4\",\n  \"nutrition\": \"Nutrition information\",\n  \"is_recommended\": true,\n  \"is_quickMenu\": false\n}"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Unauthorized",
                  "message": "Credentials not provided"
                }
              }
            }
          }
        }
      }
    },
    "/foods": {
      "get": {
        "tags": ["default"],
        "summary": "Get foods",
        "parameters": [
          {
            "name": "storeId",
            "in": "query",
            "description": "ID of the store",
            "required": false,
            "schema": {
              "type": "ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Unauthorized",
                  "message": "Credentials not provided"
                }
              }
            }
          }
        }
      },
      "components": {
        "securitySchemes": {
          "cookieAuth": {
            "type": "apiKey",
            "name": "Cookie",
            "in": "cookie",
            "description": "Enter your authentication cookie",
            "credential": "include"
          }
        }
      }
    },
    "/foods/{foodId}": {
      "put": {
        "tags": ["default"],
        "summary": "update food",
        "parameters": [
          {
            "name": "foodId",
            "in": "path",
            "description": "ID of the food",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "foodName": {
                    "foodName": "Cheese Burder"
                  }
                }
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "string",
                    "example": "{\n    \"foodName\":\"israfil\"\n}"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        },

        "401": {
          "description": "Unauthorized",
          "content": {
            "application/json": {
              "example": {
                "error": "Unauthorized",
                "message": "Credentials not provided"
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      },
      "components": {
        "securitySchemes": {
          "cookieAuth": {
            "type": "apiKey",
            "name": "Cookie",
            "in": "cookie",
            "description": "Enter your authentication cookie",
            "credential": "include"
          }
        }
      }
    },
...
  }
}